Speedup DictList.remove
This speeds up the function ~3x on iJO1366 and likely more for larger
models. This replaces list.remove(x) with list.pop(DictList.index(x))
which prevents n string replacements from being done and replaces them
with the o(1) dict lookup. This is okay because the items are unique in
the DictList