visbio maps
import visbio maps. These maps are scrollable and zoomable.
Additionally, they can be viewed in standalone html or in a notebok. A
button allows svg download of the maps.

This works as of zakandrewking/visbio@6ab047e

Eventually this should include more "helper" code to improve interop
between visbio and cobrapy, for example with id translations