sbml3.py for sbml lvl 3 + fbc IO (pure python)
Instead of using libsbml for parsing, this uses pure python. This means
that sbml i/o can occur without install the libsbml library. Because
this is pure python, the code runs in jython as well. The sbml produced
passes validation (http://sbml.org/Facilities/Validator/).

This is the first step to a "new" cobra sbml file format, which uses the
fbc package to encode upper and lower bounds.

These are the tasks which need to be accomplished:
[x] read/write sbml with fbc for reaction bounds and metabolite formulas
[x] imported model should solve properly
[x] written sbml should pass validation
[ ] gene reaction rule encoding
[ ] subsystems
[ ] other annotation