reaction.gene_reaction_rule sets gpr
Instead of calling the function add_gene_reaction_rule to set the gpr,
the gene_reaction_rule can be modified directly, and the setter
automatically will call the gpr parser. This is safer because the gpr
could previously be modified with no effect.

The gene_reaction_rule information is now stored as
_gene_reaction_rule. In order to preserve pickle compatibility, this is
now accounted for in __setstate__