gurobi_solver now has change_variable_bounds
gurobipy does not allow getting variables by index, but does by name
(for some strange reason). Therefore, this now uses the index as the
name (ugh) so we can get this functionality. Otherwise, we would need to
index the specific variable out of the entire list of lp.getVars(),
which takes nearly 500us every time.