DictList no longer uses _object_dict (and speedup)
It is faster to access an item from a list than from a dict (at least at
sizes up to ~10^6 elements), and simpler to just use the list as well.
This also means that operations have only one index to update (the
_dict) instead of 2.

Also improved a few of the list methods.