Follow SBML Rerversibility tag guidelines
SBML requires that an irreversible rection never be negative. However,
cobrapy's definition of reversibility is more consistent with the
biochemical definition that flux flows both ways. Therefore, set the
SBML reversibility flag when writing out a file by using the criteria of
lower_bound < 0 instead of the Reaction.reversibility property.

This fixes #114.