Fix cobra.io.json types and attributes
This fixes issue #98 where numpy objects failed to be saved.

The exclude_attributes functionality was also removed. This will cause
the json input and output to be consistent and reproducible, reducing
the possibility of generating malformed files. The module will export
the minimal amount of information needed to rebuild the cobra model.

In addition, the module can now load and dump json strings. The code
also passes pep8.