Fix cobra.Gene.remove_from_model
Going through the gene reaction rule setter caused all the gene
associations to be recomputed, causing an error when attempting to
remove the gene. Because the association does not need to be entirely
recomputed, setting the underlying _gene_reaction_rule instead fixes
this issue. This closes #119 on github.

In addition, the model parameter was unecessary to this function and has
been deprecated.