Add ability to compile gene_reaction_rule
For most gene reaction rules, compiling the string allows a direct eval
using a substitution dict to desribe the activation state of the genes
allows for much faster computation than using regular expressions and
string manipulations to transform the expression to a satement without
substitutions.

This enables detecting and using gene_reaction_rules which can be
compiled. This is now done by default for double gene deletions.