use container-based travis
This is faster, but means that sudo can not be used. The testing
dependencies are all downloaded. The glpk and gmp libraries are
downloaded as .a libs for static linking, and esolver is also statically
linked to gmp.

To ensure the static linking works when compiling cglpk, the LDFLAGS
need to be overridden.