Clean Species/Metabolite/Formula Relationship
The Formula class is now deprecated (fixes #146). Instead, formulas will
be stored as a string. Moreover, because a Gene can not have a formula,
the formula property has been removed from Species (fixes #145). Other
simplifications include removal of the mnx_id argument from object and
the positional attributes from Gene.

Additionally, the composition of a Metabolite is now a lazy property
(fixes #151).