units, compartments, SBO and fbc:label in sbml3
Add an option which correctly creates and assigns units.

Read in compartments from the SBML file.

The spec has been modified to have a required fbc:label on each gene,
and added the fbc:strict attribute to the Model. Also escape gene id's
with "G_".

Add sboTerm entries to the fluxBounds and Document.