replace moma implementation
Implementation is now simpler, but still passes the same unit tests.

This fixes #135 and #65.

It is worth noting that at this point in time, CPLEX sometimes fails to
find the optimal solution on difficult problems (although gurobi seems
to hav no trouble). The optimal QP settings for CPLEX still need to be
determined.