rework of double deletion
This has been refactored to be the same way as single_deletion. The
functions double_reaction_deletion and double_gene_deletion do the
setup for the problem, and will (based on the method keyword argument)
pass on the computational work to the _fba and _moma helper functions.

In the spirit of DRY, lots of the work is now done by helper functions.

MOMA for double deletion now uses the more efficient MOMA setup.

The documentation is was also really beefed up for the functions in this
section.