build windows wheel with appveyor
The appveyor continous integration service runs on Windows,
complementing travis (which runs on Ubuntu Linux). Additionally, this
will allow compiling of pre-compiled binary wheels automatically to
simplify deployment on Microsoft Windows.

This will download pre-built glpk.lib libraries for static linking. As
stated in the documentation, the source for glpk is available here:
http://www.gnu.org/software/glpk/