sbml3 is default SBML reader and writer
When an older model is detected, the function will fall back to the
older legacy function in sbml.py. Additionally, when writing SBML,
use_fbc_package can be used to write out a model in the older format.

A mini model was added in all the formats COBRA can parse. It is now
used for all of the I/O tests. Older test files were removed.

A textbook model in the fbc format was also added as a test model. This
is now used as the model for most of the tests.