core element cleanup - all have id and name
This is more consistent (Model.description instead of Model.name was
arbitrary and confusing). This is also more consistent with SBML.

Model.description is still available as a property for compatibility
reasons, but it will raise a warning when accessed.

Some deprecated methods for core classes were also finally removed.

The I/O has been updated to account for the description attribute being
replaced by name.

This fixes #175