Adding summary methods to Model and Metabolite
This commit adds methods to core.Model and core.Metabolite which print
basic statistics to describe FBA solutions. Model.summary() shows
information on the input-output behavior of the model and on objective
values, while Metabolite.summary() shows the primary in and out streams
of mass flow through the given metabolite node.