Make FVA a bit faster. (#286)
The FVA in cobrapy alternated between optimizing via minimum
and maximum, but the minimum solutions tend to be closer
to each other than the maximum solutions. Thus, a better
recycling strategy for the solution basis is to minimize
all reactions first followed by maximizing them. This is also
the algorithm described in the original publication.