refactor reaction objective_coefficient
reaction.objective_coefficient makes less sense when the objective can
be made up of more than just reactions and their coefficients. However,
we still want the option to easily assign the model.objective a set of
reactions and their coefficients.

This is an attempt to reduce this inconsistency by moving the property
reaction.objective_coefficient to model.set_reaction_coefficient and
model.reaction_coefficient where model._objective_reactions is a set
that keeps track of the reactions that have been assigned as objectives
and the coefficient being extracted from the solver.