feat: adding context managers for model manipulation (#312)
Make model as context manager using the HistoryManager class, supporting the reversing of reaction bounds and setting the objective in possibly nested contexts. Using contexts is recommended for evaluating the effect of changes to a model and then jumping back to the previous model as copying models is a computationally expensive operation.