fix: use internally consistent key w/o underscores
Underscores in keys for KEY: value associations in the notes field are
replaced with spaces so internally use only underscore free keys or
key-values may get repeated in legacy sbml output. Tricky to create
meaningful test for this as the notes field as (mis)using
legacy/incomplete sbml specification for sake of backwards
compatibility.

Fix #547.