fix: summary now accepts solution kwarg, resolves if missing (#566)
This is a PR to fix #520 and de-couple the summary methods from connection via reaction.flux to the underlying solver in memory.

model.summary() now accepts a kwarg solution, which if given will re-use the fluxes from a previous solution. Otherwise puts in a call to model.optimize() to get an updated set of fluxes. Speeds don't seem to be affected that much, especially if the solver is able to re-use the solution basis.