Improve convergence and stability of sampling.
Fixes several numerical issues with sampling. There are still
limitations though, mostly since it is hard to have a numerically
stable online calculation of the mean. Also finally fixes the
RecursionError possibility. Sampling now throws a real and informative
error if it can not get out of a small sampling region.