gapfilling.ipynb bug fix
Hi,
I noticed that the gapfilling.ipynb notebook in the documentation has an error in the second cell:
First the reaction is not called correctly. I changed add_reaction() for add_reactions() which the correct way of calling it.
Second, the argument in the add_reactions() function needs to be an iterable so I enclosed the reaction.copy() in [].
Third, the cell needs to import cobra to work.

Finally, some of the output that I got in the following cells differ a little bit from what you have in the original notebook.

I hope you find these changes useful.